requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
alembic==1.13.1
psycopg[binary]==3.2.10
line-bot-sdk==3.7.0
requests==2.31.0
APScheduler==3.10.4
certifi
urllib3
Flask-Admin==1.6.1
